# Optimize ConstantExpr::evalSpecialForm for strings