--- a/velox/vector/SimpleVector.h
+++ b/velox/vector/SimpleVector.h
@@ -286,12 +286,12 @@ class SimpleVector : public BaseVector {
   }
 
   /// Computes and saves is-ascii flag for a given set of rows if not already
-  /// present.
+  /// present. Returns computed value.
   template <typename U = T>
-  typename std::enable_if<std::is_same<U, StringView>::value, void>::type
+  typename std::enable_if<std::is_same<U, StringView>::value, bool>::type
   computeAndSetIsAscii(const SelectivityVector& rows) {
     if (rows.isSubset(asciiSetRows_)) {
-      return;
+      return isAllAscii_;
     }
     ensureIsAsciiCapacity(rows.end());
     bool isAllAscii = true;
@@ -311,6 +311,7 @@ class SimpleVector : public BaseVector {
     }
 
     asciiSetRows_.select(rows);
+    return isAllAscii_;
   }
 
   /// Clears asciiness state.