@funrollloops 2021-10-29
Looks like folly::Benchmark has some issues with timer precision at small values?
```
in_int(spark_rhs1)                                  inf%     0.00fs  Infinity
```
I'm confused about how the PrestoSQL version of IN is faster than this version for the N=1 case, _even after I added a special case for it that just uses ==!_ What in the world is it doing?